+++ src/aliases.rs
@@ -1,8 +1,8 @@
 use nalgebra as na;
 use rand;
 
-pub type Vec3 = na::Vector3<f64>;
-pub type Vec2 = na::Vector2<f64>;
-pub type Mat4 = na::Matrix4<f64>;
-pub type Mat3 = na::Matrix3<f64>;
+pub type Vec3 = na::Vector3<f32>;
+pub type Vec2 = na::Vector2<f32>;
+pub type Mat4 = na::Matrix4<f32>;
+pub type Mat3 = na::Matrix3<f32>;
 pub type RandGen = rand::prelude::ThreadRng;
