+++ src/hitable/bvh_node.rs
@@ -18,7 +18,7 @@ impl BvhNode {
     /// Constructor
     /// time_0, time_1: used for moving hitables.
     /// list: each element should have non-None bounding box
-    pub fn new(mut list: Vec<Arc<Hitable>>, time_0: f64, time_1: f64) -> Self {
+    pub fn new(mut list: Vec<Arc<Hitable>>, time_0: f32, time_1: f32) -> Self {
         if list.is_empty() {
             return BvhNode {
                 left: Arc::new(Empty::new()),
@@ -69,7 +69,7 @@ impl BvhNode {
     ) -> (usize, usize) {
         let mut min_axis = 0;
         let mut min_idx = 0;
-        let mut min_cost = std::f64::MAX;
+        let mut min_cost = std::f32::MAX;
         for axis in 0..3 {
             bboxes.sort_unstable_by(|a, b| a.compare_center(b, axis));
             let areas = Self::calc_consecutive_bboxes_united_areas(&bboxes);
@@ -91,8 +91,8 @@ impl BvhNode {
     /// For example, if list = [a,b,c], then it returns
     /// [(|0|,|a+b+c|), (|a|,|b+c|), (|a+b|,|c|), (|a+b+c|,|0|)]
     /// where + is the unite operator, and |x| is the surface area of x.
-    fn calc_consecutive_bboxes_united_areas(list: &Vec<Aabb>) -> Vec<(f64, f64)> {
-        let mut areas = Vec::<(f64, f64)>::new();
+    fn calc_consecutive_bboxes_united_areas(list: &Vec<Aabb>) -> Vec<(f32, f32)> {
+        let mut areas = Vec::<(f32, f32)>::new();
         areas.resize(list.len() + 1, (0.0, 0.0));
         let mut bbox_f = Aabb::empty();
         let mut bbox_b = Aabb::empty();
@@ -106,7 +106,7 @@ impl BvhNode {
         }
         areas
     }
-    fn sort_hitables_center(list: &mut Vec<Arc<Hitable>>, axis: usize, time_0: f64, time_1: f64) {
+    fn sort_hitables_center(list: &mut Vec<Arc<Hitable>>, axis: usize, time_0: f32, time_1: f32) {
         list.sort_unstable_by(|a, b| {
             let a_box = a.bounding_box(time_0, time_1).unwrap(); // expect panic if None
             let b_box = b.bounding_box(time_0, time_1).unwrap();
@@ -115,13 +115,13 @@ impl BvhNode {
     }
     /// Cost of traversing into a group according to SAH = Surface Area Heuristics, i.e.,
     /// (surface area of bbox) * (# of primitives)
-    fn cost_sah(box_surface_area: f64, primitives: i32) -> f64 {
-        box_surface_area * primitives as f64
+    fn cost_sah(box_surface_area: f32, primitives: i32) -> f32 {
+        box_surface_area * primitives as f32
     }
 }
 
 impl Hitable for BvhNode {
-    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f64, t_max: f64) -> Option<HitRecord<'s>> {
+    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f32, t_max: f32) -> Option<HitRecord<'s>> {
         if !self.aabb.hit(ray, t_min, t_max) {
             return None;
         }
@@ -141,7 +141,7 @@ impl Hitable for BvhNode {
             second.hit(ray, t_min, t_max)
         }
     }
-    fn bounding_box(&self, _t0: f64, _t1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _t0: f32, _t1: f32) -> Option<Aabb> {
         Some(self.aabb)
     }
 }
