+++ src/hitable/sphere.rs
@@ -7,17 +7,17 @@ use crate::onb::Onb;
 use crate::pdf::random_in_cone;
 use crate::ray::Ray;
 use nalgebra as na;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use std::sync::Arc;
 
 pub struct Sphere {
     center: Vec3,
-    radius: f64,
+    radius: f32,
     material: Arc<Material>,
 }
 
 impl Sphere {
-    pub fn new(center: &Vec3, radius: f64, material: Arc<Material>) -> Self {
+    pub fn new(center: &Vec3, radius: f32, material: Arc<Material>) -> Self {
         Sphere {
             center: *center,
             radius: radius,
@@ -25,17 +25,17 @@ impl Sphere {
         }
     }
     /// Calculates the parameter t of the ray at which it hits this Sphere.
-    pub fn hit_core(center: &Vec3, radius: f64, ray: &Ray, t_min: f64, t_max: f64) -> Option<f64> {
+    pub fn hit_core(center: &Vec3, radius: f32, ray: &Ray, t_min: f32, t_max: f32) -> Option<f32> {
         let oc = ray.origin - center;
         let a = na::dot(&ray.direction, &ray.direction);
         let b = na::dot(&oc, &ray.direction);
         let c = na::dot(&oc, &oc) - radius * radius;
-        let disc: f64 = b * b - a * c;
+        let disc: f32 = b * b - a * c;
         if disc <= 0.0 {
             return None;
         }
-        let disc_rt = f64::sqrt(disc);
-        let mut t: f64;
+        let disc_rt = f32::sqrt(disc);
+        let mut t: f32;
         t = (-b - disc_rt) / a;
         if t_min < t && t < t_max {
             return Some(t);
@@ -48,14 +48,14 @@ impl Sphere {
     }
     /// convert a point on unit sphere to a uv coordinate
     pub fn get_uv(p: &Vec3) -> Vec2 {
-        let phi = f64::atan2(p[2], p[0]);
-        let theta = f64::asin(p[1].min(1.0).max(-1.0));
+        let phi = f32::atan2(p[2], p[0]);
+        let theta = f32::asin(p[1].min(1.0).max(-1.0));
         Vec2::new(0.5 - 0.5 * (phi / PI), 0.5 + theta / PI)
     }
 }
 
 impl Hitable for Sphere {
-    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f64, t_max: f64) -> Option<HitRecord<'s>> {
+    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f32, t_max: f32) -> Option<HitRecord<'s>> {
         Sphere::hit_core(&self.center, self.radius, ray, t_min, t_max).map(|t| {
             let point = ray.evaluate(t);
             let normal = (point - self.center) / self.radius;
@@ -69,7 +69,7 @@ impl Hitable for Sphere {
             }
         })
     }
-    fn bounding_box(&self, _time_0: f64, _time_1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time_0: f32, _time_1: f32) -> Option<Aabb> {
         let rad_vec = Vec3::new(self.radius, self.radius, self.radius);
         return Some(Aabb::new(
             &(self.center - rad_vec),
@@ -82,7 +82,7 @@ impl Hitable for Sphere {
         let onb = Onb::build_from_w(&(self.center - origin));
         onb.local_to_global_vec(&vec_local)
     }
-    fn direction_density(&self, origin: &Vec3, dir: &Vec3) -> f64 {
+    fn direction_density(&self, origin: &Vec3, dir: &Vec3) -> f32 {
         let cosine = (self.center - origin).normalize().dot(&dir.normalize());
         let cosine_max =
             (1.0 - self.radius.powf(2.0) / (origin - self.center).norm_squared()).sqrt();
@@ -97,12 +97,12 @@ impl Hitable for Sphere {
 pub struct MovingSphere {
     center_0: Vec3, // center at time = 0
     center_1: Vec3, // center at time = 1
-    radius: f64,
+    radius: f32,
     material: Box<Material>,
 }
 
 impl MovingSphere {
-    pub fn new(center_0: &Vec3, center_1: &Vec3, radius: f64, material: Box<Material>) -> Self {
+    pub fn new(center_0: &Vec3, center_1: &Vec3, radius: f32, material: Box<Material>) -> Self {
         MovingSphere {
             center_0: *center_0,
             center_1: *center_1,
@@ -113,10 +113,10 @@ impl MovingSphere {
 }
 
 impl MovingSphere {
-    pub fn center_at(&self, time: f64) -> Vec3 {
+    pub fn center_at(&self, time: f32) -> Vec3 {
         self.center_0 + time * (self.center_1 - self.center_0)
     }
-    pub fn bounding_box_at(&self, time: f64) -> Aabb {
+    pub fn bounding_box_at(&self, time: f32) -> Aabb {
         let rad_vec = Vec3::new(self.radius, self.radius, self.radius);
         let center = self.center_at(time);
         Aabb::new(&(center - rad_vec), &(center + rad_vec))
@@ -124,7 +124,7 @@ impl MovingSphere {
 }
 
 impl Hitable for MovingSphere {
-    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f64, t_max: f64) -> Option<HitRecord<'s>> {
+    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f32, t_max: f32) -> Option<HitRecord<'s>> {
         let center = self.center_at(ray.time);
         Sphere::hit_core(&center, self.radius, ray, t_min, t_max).map(|t| {
             let point = ray.evaluate(t);
@@ -139,7 +139,7 @@ impl Hitable for MovingSphere {
             }
         })
     }
-    fn bounding_box(&self, time_0: f64, time_1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, time_0: f32, time_1: f32) -> Option<Aabb> {
         Some(Aabb::unite(
             &self.bounding_box_at(time_0),
             &self.bounding_box_at(time_1),
