+++ src/hitable/triangle.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 pub struct Triangle {
     vertices: [Vec3; 3],              // a,b,c
     perpendicular_normals: [Vec3; 3], // (aから辺bcへの単位垂線ベクトル, bから辺caへの..., cから辺abへの...)
-    perpendicular_lengths: [f64; 3],  // 垂線の長さ
+    perpendicular_lengths: [f32; 3],  // 垂線の長さ
     normal: Vec3,                     // 単位法線ベクトル。(b-a).cross(c-a).normalize
     material: Arc<Material>,
     vertex_normals: Option<[Vec3; 3]>, // Normal vectors at vertices. If this is Some, the hit() returns HitRecord with linearly interpolated normal vector.
@@ -44,7 +44,7 @@ impl Triangle {
 }
 
 impl Hitable for Triangle {
-    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f64, t_max: f64) -> Option<HitRecord<'s>> {
+    fn hit<'s, 'r>(&'s self, ray: &'r Ray, t_min: f32, t_max: f32) -> Option<HitRecord<'s>> {
         let denom = self.normal.dot(&ray.direction);
         if denom == 0.0 {
             return None;
@@ -54,7 +54,7 @@ impl Hitable for Triangle {
             return None;
         }
         let p = ray.evaluate(t);
-        let mut weights: [f64; 3] = [0.0, 0.0, 0.0];
+        let mut weights: [f32; 3] = [0.0, 0.0, 0.0];
         for i in 0..3 {
             weights[i] = 1.0
                 - (p - self.vertices[i]).dot(&self.perpendicular_normals[i])
@@ -79,7 +79,7 @@ impl Hitable for Triangle {
             material: self.material.as_ref(),
         })
     }
-    fn bounding_box(&self, _time_0: f64, _time_1: f64) -> Option<Aabb> {
+    fn bounding_box(&self, _time_0: f32, _time_1: f32) -> Option<Aabb> {
         Some(Aabb::from_points(vec![
             self.vertices[0],
             self.vertices[1],
@@ -89,7 +89,7 @@ impl Hitable for Triangle {
     fn random_direction_from(&self, _origin: &Vec3, _rng: &mut RandGen) -> Vec3 {
         unimplemented!()
     }
-    fn direction_density(&self, _origin: &Vec3, _dir: &Vec3) -> f64 {
+    fn direction_density(&self, _origin: &Vec3, _dir: &Vec3) -> f32 {
         unimplemented!()
     }
 }
