+++ src/material/mod.rs
@@ -19,5 +19,5 @@ pub trait Material: Send + Sync {
     /// A characteristic function of scattering process.
     /// Propotional to the reciprocal of BRDF.
     /// This function must be normalized as a pdf in the "scattered" argument.
-    fn scattering_pdf(&self, ray: &Ray, scattered: &Ray, rec: &HitRecord) -> f64;
+    fn scattering_pdf(&self, ray: &Ray, scattered: &Ray, rec: &HitRecord) -> f32;
 }
